+++ include/generate/decorator/tree.h
@@ -110,7 +110,7 @@ public:
     virtual std::shared_ptr<const DecoratorInstance> createInstance(PositionI chunkBasePosition, PositionI columnBasePosition, PositionI surfacePosition,
                                  WorldLockManager &lock_manager, BlockIterator chunkBaseIterator,
                                  const checked_array<checked_array<checked_array<Block, BlockChunk::chunkSizeZ>, BlockChunk::chunkSizeY>, BlockChunk::chunkSizeX> &blocks,
-                                 RandomSource &randomSource, std::size_t generateNumber) const override
+                                 RandomSource &randomSource, std::uint32_t generateNumber) const override
     {
         VectorI surfaceRPos = surfacePosition - chunkBasePosition;
         Block groundBlock = blocks[surfaceRPos.x][surfaceRPos.y][surfaceRPos.z];
